# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cadife Smart Travel API",
    "description": "Backend inteligente para turismo via WhatsApp + Flutter. Orquestra webhooks da Meta, processamento de IA (RAG + LangChain), gestão de leads, propostas e agendamentos. Documentação completa disponível em /docs (Swagger UI) e /redoc (ReDoc).",
    "contact": {
      "name": "Cadife Tour - Time de Desenvolvimento",
      "url": "https://cadifetour.com.br/",
      "email": "dev@cadifetour.com.br"
    },
    "license": {
      "name": "Confidencial — Uso Interno do Time de Desenvolvimento"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/leads": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "Listar leads",
        "description": "Retorna leads paginados com suporte a filtros por status, score e busca textual. Consultores visualizam apenas seus próprios leads; admin e agência visualizam todos.",
        "operationId": "list_leads_leads_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro por status do lead",
              "title": "Status"
            },
            "description": "Filtro por status do lead"
          },
          {
            "name": "score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro por score (quente, morno, frio)",
              "title": "Score"
            },
            "description": "Filtro por score (quente, morno, frio)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Busca textual em nome ou telefone",
              "title": "Search"
            },
            "description": "Busca textual em nome ou telefone"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Número da página",
              "default": 1,
              "title": "Page"
            },
            "description": "Número da página"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Itens por página (máx. 100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Itens por página (máx. 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadListResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Perfil sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Leads"
        ],
        "summary": "Criar lead",
        "description": "Cria um novo lead manualmente. Caso o telefone já exista, retorna 409 Conflict.",
        "operationId": "create_lead_leads_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetailDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Perfil sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação no body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Telefone já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/leads/my-active": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "Lead ativo do cliente logado",
        "description": "Retorna o lead associado ao telefone do usuário autenticado (perfil cliente). Se não existir, cria automaticamente um novo lead.",
        "operationId": "get_my_active_lead_leads_my_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetailDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Usuário sem telefone cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/leads/manual": {
      "post": {
        "tags": [
          "Leads"
        ],
        "summary": "Create Manual Lead",
        "description": "Criação manual de lead via App Agência.\nPermite preenchimento de briefing inicial e atribuição de consultor.",
        "operationId": "create_manual_lead_leads_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualLeadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetailDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/leads/{lead_id}/interacoes": {
      "get": {
        "tags": [
          "Leads"
        ],
        "summary": "Histórico de interações do lead",
        "description": "Retorna todas as mensagens trocadas entre o cliente e a IA para um lead específico.",
        "operationId": "get_interacoes_leads__lead_id__interacoes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteracaoListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão para este lead",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Lead não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HTTPErrorResponse",
        "description": "Generic error envelope returned by all failed requests.",
        "example": {
          "detail": "Descrição legível do erro",
          "error_code": "ERR_EXAMPLE"
        }
      },
      "HTTPValidationErrorResponse": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HTTPValidationErrorResponse",
        "description": "Pydantic / FastAPI validation error (422).",
        "example": {
          "detail": [
            {
              "loc": [
                "body",
                "telefone"
              ],
              "msg": "field required",
              "type": "missing"
            }
          ]
        }
      },
      "InteracaoListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InteracaoResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "InteracaoListResponse"
      },
...
      "LeadCreateRequest": {
        "properties": {
          "nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "telefone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 8,
            "title": "Telefone"
          },
          "origem": {
            "$ref": "#/components/schemas/LeadOrigem",
            "default": "whatsapp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "telefone"
        ],
        "title": "LeadCreateRequest"
      },
...
      "LeadListResponseDTO": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LeadListItemDTO"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "pages"
        ],
        "title": "LeadListResponseDTO"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}